backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.0.4
psycopg2-binary>=2.8
requests==2.25.1
uvicorn==0.13.4
whitenoise==5.2.0
drf-spectacular~=0.21.0
dj-rest-auth==2.1.11
djangorestframework-simplejwt==5.2.0
django-allauth~=0.48.0